Change implicit conversion of IntAnswer to DecAnswer now that DecAnswer takes a BigDecimal

+++ niro/src/main/scala/nl/uva/se/sc/niro/model/Expressions/ImplicitConversions.scala
@@ -5,5 +5,5 @@ import nl.uva.se.sc.niro.model.Expressions.answers.{ DecAnswer, IntAnswer }
 import scala.language.implicitConversions
 
 object ImplicitConversions {
-  implicit def intAnswerToDecAnswer(value: IntAnswer): DecAnswer = DecAnswer(value.possibleValue.map(_.toDouble))
+  implicit def intAnswerToDecAnswer(value: IntAnswer): DecAnswer = DecAnswer(value.possibleValue.map(BigDecimal(_)))
 }
